+++ src/app/providers/UserProvider/UserProvider.tsx
@@ -1,12 +1,12 @@
-import { type JSX, type ProviderProps, useReducer } from 'react'
-import { UserActionType, userContext, type UserState } from './UserContext'
-import { useFetch } from '@/shared/hooks/useFetch.tsx'
-import type { ClientResponse, Customer } from '@commercetools/platform-sdk'
-import { api } from '@/server/api.ts'
-import { isAnonymous } from '@/server/client.ts'
-import Loader from '@/shared/ui/Loader/Loader.tsx'
-import { userReducer } from '@/app/providers/UserProvider/UserReducer.ts'
+import { api } from '@/server/api'
 import { toast } from 'react-hot-toast'
+import { isAnonymous } from '@/server/client'
+import Loader from '@/shared/ui/Loader/Loader'
+import { useFetch } from '@/shared/hooks/useFetch'
+import { userContext, type UserState } from './UserContext'
+import { UserActionType, userReducer } from './UserReducer'
+import { type JSX, type ProviderProps, useReducer } from 'react'
+import type { ClientResponse, Customer } from '@commercetools/platform-sdk'
 
 export const UserProvider = ({ value, children }: Partial<ProviderProps<UserState>>): JSX.Element => {
   const [state, dispatch] = useReducer(userReducer, value || {})
